requirements.txt

streamlit>=1.22.0
langchain>=0.0.267
langchain-community>=0.0.16
openai>=1.2.0
yt-dlp>=2023.10.13
pydub>=0.25.1
python-dotenv>=1.0.0
chromadb>=0.4.15
langchain-openai>=0.0.2
ffmpeg-python>=0.2.0
aiohttp>=3.8.5